Unify context handling between threads, events, and async tasks
